Add capture to webhook and websocket
