+++ frontend/src/lib/components/triggers/WebsocketTriggerEditorInner.svelte
@@ -24,8 +24,7 @@
 	import { fade } from 'svelte/transition'
 	import JsonEditor from '../apps/editor/settingsPanel/inputEditor/JsonEditor.svelte'
 	import type { Schema } from '$lib/common'
-	import ToggleButtonGroup from '../common/toggleButton-v2/ToggleButtonGroup.svelte'
-	import ToggleButton from '../common/toggleButton-v2/ToggleButton.svelte'
+	import WebsocketEditorConfigSection from './WebsocketEditorConfigSection.svelte'
 
 	let drawer: Drawer
 	let is_flow: boolean = false
@@ -316,92 +315,14 @@
 					</Label>
 				</div>
 
-				<Section label="Websocket">
-					<div class="mb-2">
-						<ToggleButtonGroup
-							selected={url.startsWith('$') ? 'runnable' : 'static'}
-							on:selected={(ev) => {
-								url = ev.detail === 'runnable' ? '$script:' : ''
-								url_runnable_args = {}
-							}}
-						>
-							<ToggleButton value="static" label="Static URL" />
-							<ToggleButton value="runnable" label="Runnable result as URL" />
-						</ToggleButtonGroup>
-					</div>
-					{#if url.startsWith('$')}
-						<div class="flex flex-col w-full gap-4">
-							<div class="block grow w-full">
-								<div class="text-secondary text-sm flex items-center gap-1 w-full justify-between">
-									<div>
-										Runnable
-										<Required required={true} />
-									</div>
-								</div>
-								<ScriptPicker
-									allowFlow={true}
-									itemKind={url.startsWith('$flow:') ? 'flow' : 'script'}
-									initialPath={url.split(':')[1] ?? ''}
-									on:select={(ev) => {
-										dirtyUrl = true
-										const { path, itemKind } = ev.detail
-										url = `$${itemKind}:${path ?? ''}`
-									}}
-								/>
-								<div class="text-red-600 dark:text-red-400 text-2xs mt-1.5">
-									{dirtyUrl ? urlError : ''}
-								</div>
-							</div>
-						</div>
-
-						{#if url.split(':')[1]?.length > 0}
-							{#if urlRunnableSchema}
-								<p class="font-semibold text-sm mt-4 mb-2">Arguments</p>
-								{#await import('$lib/components/SchemaForm.svelte')}
-									<Loader2 class="animate-spin mt-2" />
-								{:then Module}
-									<Module.default
-										schema={urlRunnableSchema}
-										bind:args={url_runnable_args}
-										shouldHideNoInputs
-										class="text-xs"
-									/>
-								{/await}
-								{#if urlRunnableSchema.properties && Object.keys(urlRunnableSchema.properties).length === 0}
-									<div class="text-xs texg-gray-700">This runnable takes no arguments</div>
-								{/if}
-							{:else}
-								<Loader2 class="animate-spin mt-2" />
-							{/if}
-						{/if}
-					{:else}
-						<div class="flex flex-col w-full gap-4">
-							<label class="block grow w-full">
-								<div class="text-secondary text-sm flex items-center gap-1 w-full justify-between">
-									<div>
-										URL
-										<Required required={true} />
-									</div>
-								</div>
-								<input
-									type="text"
-									autocomplete="off"
-									bind:value={url}
-									disabled={!can_write}
-									on:input={() => {
-										dirtyUrl = true
-									}}
-									class={urlError === ''
-										? ''
-										: 'border border-red-700 bg-red-100 border-opacity-30 focus:border-red-700 focus:border-opacity-30 focus-visible:ring-red-700 focus-visible:ring-opacity-25 focus-visible:border-red-700'}
-								/>
-								<div class="text-red-600 dark:text-red-400 text-2xs mt-1.5">
-									{dirtyUrl ? urlError : ''}
-								</div>
-							</label>
-						</div>
-					{/if}
-				</Section>
+				<WebsocketEditorConfigSection
+					{url}
+					{url_runnable_args}
+					{dirtyUrl}
+					{urlError}
+					{urlRunnableSchema}
+					{can_write}
+				/>
 
 				<Section label="Runnable">
 					<p class="text-xs mb-1 text-tertiary">
